+++ Sources/LearningImpl/src/bagaturchess/deeplearning/impl/eval/pst/NeuralNetworkEvaluator.java
@@ -32,19 +32,19 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase1() {
+	protected int phase1() {
 		
 		NeuralNetworkUtils.clearInputsArray(inputs);
 		NeuralNetworkUtils_PST.fillInputs(network, inputs, bitboard);
 		NeuralNetworkUtils.calculate(network);
 		double actualWhitePlayerEval = NeuralNetworkUtils.getOutput(network);
 
-		return actualWhitePlayerEval;
+		return (int) actualWhitePlayerEval;
 	}
 	
 	
 	@Override
-	protected double phase2() {
+	protected int phase2() {
 
 		int eval = 0;
 		
@@ -53,7 +53,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase3() {
+	protected int phase3() {
 		
 		int eval = 0;
 				
@@ -62,7 +62,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase4() {
+	protected int phase4() {
 		
 		int eval = 0;
 		
@@ -71,7 +71,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase5() {
+	protected int phase5() {
 		
 		int eval = 0;
